refactoring functions into smaller files
