+++ src/clientApp.ts
@@ -7,62 +7,47 @@ import JSZip from 'jszip';
 import Koa from 'koa';
 import views from '@ladjs/koa-views';
 import koaConditionalGet from 'koa-conditional-get';
-import fetch from 'node-fetch';
 import path from 'path';
 import { Transform } from 'stream';
 import { fileURLToPath, URL } from 'url';
 import chalk from 'chalk';
-import { getScreepsPath } from './utils/steamGamePath';
-import { removeRoomDecorations, clientStartup, generateScriptTag, modifyClientAppMenu } from './utils/clientScripts';
-import { Server } from './utils/types';
 import { Client, Route } from './utils/client';
+import { getScreepsPath } from './utils/gamePath';
+import {
+    logError,
+    isOfficialLikeVersion,
+    trimLocalSubdomain,
+    generateScriptTag,
+    getServerListConfig,
+    extractBackend,
+} from './utils/clientUtils';
+import { clientAuth } from './inject/clientAuth';
+import { removeDecorations } from './inject/removeDecorations';
+import { customMenuLinks } from './inject/customMenuLinks';
 
 // Log welcome message
 console.log('🧩', chalk.yellowBright('Screepers Steamless Client'));
 
 // Parse program arguments
-const argv = (function () {
+const argv = (() => {
     const parser = new ArgumentParser();
-    parser.add_argument('--beautify', {
-        action: 'store_true',
-        default: false,
-    });
-    parser.add_argument('--package', {
-        nargs: '?',
-        type: 'str',
-    });
-    parser.add_argument('--port', {
-        nargs: '?',
-        type: 'int',
-    });
-    parser.add_argument('--backend', {
-        nargs: '?',
-        type: 'str',
-    });
-    parser.add_argument('--host', {
-        nargs: '?',
-        type: 'str',
-    });
-    parser.add_argument('--internal_backend', {
-        nargs: '?',
-        type: 'str',
-    });
-    parser.add_argument('--server_list', {
-        nargs: '?',
-        type: 'str',
-    });
+    parser.add_argument('--beautify', { action: 'store_true', default: false });
+    parser.add_argument('--package', { nargs: '?', type: 'str' });
+    parser.add_argument('--port', { nargs: '?', type: 'int' });
+    parser.add_argument('--backend', { nargs: '?', type: 'str' });
+    parser.add_argument('--host', { nargs: '?', type: 'str' });
+    parser.add_argument('--internal_backend', { nargs: '?', type: 'str' });
+    parser.add_argument('--server_list', { nargs: '?', type: 'str' });
     return parser.parse_args();
 })();
 
-export const error = (...args: unknown[]) => console.error('❌', chalk.bold.red('Error'), ...args);
-
 // Create proxy
 const proxy = httpProxy.createProxyServer({ changeOrigin: true });
-proxy.on('error', (err) => error(err));
+proxy.on('error', (err) => logError(err));
 
 const exitOnPackageError = () => {
-    error('Could not find the Screeps "package.nw".');
-    error('Use the "--package" argument to specify the path to the "package.nw" file.');
+    logError('Could not find the Screeps "package.nw".');
+    logError('Use the "--package" argument to specify the path to the "package.nw" file.');
     process.exit(1);
 };
 
@@ -83,70 +68,18 @@ await zip.loadAsync(data);
 // HTTP header is only accurate to the minute
 const lastModified = stat.mtime;
 
-// Set up koa server
+// Set up web server
 const koa = new Koa();
 const port = argv.port ?? 8080;
 const host = argv.host ?? 'localhost';
 const server = koa.listen(port, host);
-server.on('error', (err) => error(err));
-
-// Extract backend and endpoint from URL
-const extract = (url: string) => {
-    if (argv.backend) {
-        return {
-            backend: argv.backend.replace(/\/+$/, ''),
-            endpoint: url,
-        };
-    }
-    const groups = /^\/\((?<backend>[^)]+)\)(?<endpoint>\/.*)$/.exec(url)?.groups;
-    if (groups) {
-        return {
-            backend: groups.backend.replace(/\/+$/, ''),
-            endpoint: groups.endpoint,
-        };
-    }
-};
+server.on('error', (err) => logError(err));
 
 // Get system path for public files dir
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 const rootDir = path.join(__dirname, '..');
 const indexFile = 'index.ejs';
-
-const getServerListConfig = async () => {
-    let serverListPath = argv.server_list;
-    if (!serverListPath) {
-        const serverListFile = 'server_list.json';
-        serverListPath = path.join(__dirname, `../settings/${serverListFile}`);
-        if (!existsSync(serverListPath)) {
-            serverListPath = path.join(__dirname, serverListFile);
-        }
-    }
-
-    const serverConfig: Server[] = JSON.parse(await fs.readFile(serverListPath, 'utf-8'));
-    const serverTypes = Array.from(new Set(serverConfig.map((server) => server.type)));
-    const serverList = serverTypes.map((type) => {
-        const serversOfType = serverConfig
-            .filter((server) => server.type === type)
-            .map((server) => {
-                const subdomain = host === 'localhost' && server.subdomain ? `${server.subdomain}.` : '';
-                const { origin, pathname } = new URL(server.url);
-                const urlpath = pathname.endsWith('/') ? pathname : `${pathname}/`;
-
-                const url = `http://${subdomain}${host}:${port}/(${origin})${urlpath}`;
-                const api = `http://${host}:${port}/(${origin})${urlpath}api/version`;
-                return { ...server, url, api };
-            });
-
-        return {
-            name: type.charAt(0).toUpperCase() + type.slice(1),
-            logo: type === 'official' ? `http://${host}:${port}/(file)/logotype.svg` : undefined,
-            servers: serversOfType,
-        };
-    });
-
-    return serverList;
-};
 
 // Setup views for rendering ejs files
 koa.use(views(path.join(__dirname, '../views'), { extension: 'ejs' }));
@@ -159,7 +92,7 @@ koa.use(async (context, next) => {
     if (argv.backend) return next(); // Skip if backend is specified
 
     if (['/', 'index.html'].includes(context.path)) {
-        const serverList = await getServerListConfig();
+        const serverList = await getServerListConfig(host, port, argv.server_list);
         if (serverList.length) {
             await context.render(indexFile, { serverList });
             return;
@@ -194,9 +127,9 @@ koa.use(async (context, next) => {
 
 // Serve client assets
 koa.use(async (context, next) => {
-    const info = extract(context.path);
+    const info = extractBackend(context.path, argv.backend);
     if (!info) {
-        error('Unknown URL', chalk.dim(context.path));
+        logError('Unknown URL', chalk.dim(context.path));
         return;
     }
 
@@ -229,137 +162,115 @@ koa.use(async (context, next) => {
     // Rewrite various payloads
     context.body = await (async function () {
         if (urlPath === 'index.html') {
-            let text = await file.async('text');
+            let src = await file.async('text');
 
+            // Client app menu links
             const seasonLink =
                 isOfficial && !prefix
                     ? `${client.getURL(Route.ROOT)}season/`
                     : client.getURL(Route.ROOT, { prefix: false });
-
             const ptrLink = isOfficial && !prefix ? `${client.getURL(Route.ROOT)}ptr/` : undefined;
-
-            function removeSubdomainFromLocalhost(host: string): string {
-                const parts = host.split('.');
-                const localhostIndex = parts.findIndex((p) => p.includes('localhost'));
-                if (localhostIndex !== -1) {
-                    host = parts[localhostIndex];
-                }
-                return host;
-            }
-
-            const serverListLink = argv.backend ? undefined : `http://${removeSubdomainFromLocalhost(clientHost)}/`;
+            const serverListLink = argv.backend ? undefined : `http://${trimLocalSubdomain(clientHost)}/`;
 
             // Inject startup script
             const header = '<title>Screeps</title>';
             const replaceHeader = [
                 header,
-                generateScriptTag(clientStartup, { backend: info.backend }),
-                generateScriptTag(removeRoomDecorations, { backend: info.backend }),
-                generateScriptTag(modifyClientAppMenu, { backend: info.backend, seasonLink, ptrLink, serverListLink }),
+                generateScriptTag(clientAuth, { backend: info.backend }),
+                generateScriptTag(removeDecorations, { backend: info.backend }),
+                generateScriptTag(customMenuLinks, { backend: info.backend, seasonLink, ptrLink, serverListLink }),
             ].join('\n');
-            text = text.replace(header, replaceHeader);
+            src = src.replace(header, replaceHeader);
 
             // Remove tracking pixels
-            text = text.replace(
+            src = src.replace(
                 /<script[^>]*>[^>]*xsolla[^>]*<\/script>/g,
                 '<script>xnt = new Proxy(() => xnt, { get: () => xnt })</script>',
             );
-            text = text.replace(
+            src = src.replace(
                 /<script[^>]*>[^>]*facebook[^>]*<\/script>/g,
                 '<script>fbq = new Proxy(() => fbq, { get: () => fbq })</script>',
             );
-            text = text.replace(
+            src = src.replace(
                 /<script[^>]*>[^>]*google[^>]*<\/script>/g,
                 '<script>ga = new Proxy(() => ga, { get: () => ga })</script>',
             );
-            text = text.replace(
+            src = src.replace(
                 /<script[^>]*>[^>]*mxpnl[^>]*<\/script>/g,
                 '<script>mixpanel = new Proxy(() => mixpanel, { get: () => mixpanel })</script>',
             );
-            text = text.replace(
+            src = src.replace(
                 /<script[^>]*>[^>]*twttr[^>]*<\/script>/g,
                 '<script>twttr = new Proxy(() => twttr, { get: () => twttr })</script>',
             );
-            text = text.replace(
+            src = src.replace(
                 /<script[^>]*>[^>]*onRecaptchaLoad[^>]*<\/script>/g,
                 '<script>function onRecaptchaLoad(){}</script>',
             );
-            return text;
+            return src;
         } else if (urlPath === 'config.js') {
-            let text = await file.async('text');
+            let src = await file.async('text');
             const opts = { full: true };
 
             // Replace API_URL, HISTORY_URL, WEBSOCKET_URL, and PREFIX in the server config
             const apiPath = client.getPath(Route.API, opts);
-            text = text.replace(/(API_URL = ')[^']*/, `$1${apiPath}/`);
+            src = src.replace(/(API_URL = ')[^']*/, `$1${apiPath}/`);
 
             const historyPath = client.getPath(Route.HISTORY, opts);
-            text = text.replace(/(HISTORY_URL = ')[^']*/, `$1${historyPath}/`);
+            src = src.replace(/(HISTORY_URL = ')[^']*/, `$1${historyPath}/`);
 
             const socketPath = client.getPath(Route.SOCKET, opts);
-            text = text.replace(/(WEBSOCKET_URL = ')[^']*/, `$1${socketPath}/`);
+            src = src.replace(/(WEBSOCKET_URL = ')[^']*/, `$1${socketPath}/`);
 
             const prefixValue = prefix?.substring(1) || '';
-            text = text.replace(/(PREFIX: ')[^']*/, `$1${prefixValue}`);
+            src = src.replace(/(PREFIX: ')[^']*/, `$1${prefixValue}`);
 
             const ptrValue = prefix === '/ptr' ? 'true' : 'false';
-            text = text.replace(/(PTR: )[^,]*/, `$1${ptrValue}`);
+            src = src.replace(/(PTR: )[^,]*/, `$1${ptrValue}`);
 
-            return text;
+            return src;
         } else if (context.path.endsWith('.js')) {
-            let text = await file.async('text');
+            let src = await file.async('text');
             if (urlPath === 'build.min.js') {
                 // Load backend info from underlying server
                 const backend = new URL(info.backend);
-                const version = await (async function () {
-                    try {
-                        const versionUrl = `${client.getURL(Route.API, { backend: true })}/version`;
-                        const response = await fetch(versionUrl);
-                        return JSON.parse(await response.text());
-                    } catch (err) {}
-                })();
-                const officialLike: boolean =
-                    version?.serverData?.features?.some(
-                        (f: { name: string }) => f.name.toLowerCase() === 'official-like',
-                    ) ?? false;
-                const official = backend.hostname === 'screeps.com' || officialLike;
-
+                const isOfficialLike = isOfficial || (await isOfficialLikeVersion(client));
                 // Look for server options payload in build information
-                for (const match of text.matchAll(/\boptions=\{/g)) {
-                    for (let i = match.index!; i < text.length; ++i) {
-                        if (text.charAt(i) === '}') {
+                for (const match of src.matchAll(/\boptions=\{/g)) {
+                    for (let i = match.index!; i < src.length; ++i) {
+                        if (src.charAt(i) === '}') {
                             try {
-                                const payload = text.substring(match.index!, i + 1);
+                                const payload = src.substring(match.index!, i + 1);
                                 // eslint-disable-next-line @typescript-eslint/no-unused-vars
                                 const holder = new Function(payload);
                                 if (payload.includes('apiUrl')) {
                                     // Inject `host`, `port`, and `official`
-                                    text = `${text.substring(0, i)},
+                                    src = `${src.substring(0, i)},
                                         host: ${JSON.stringify(backend.hostname)},
-                                        port: ${backend.port || '80'},
-                                        official: ${official},
-                                    } ${text.substring(i + 1)}`;
+                                        port: ${backend.port || '8080'},
+                                        official: ${isOfficial || isOfficialLike},
+                                    } ${src.substring(i + 1)}`;
                                 }
                                 break;
                             } catch (err) {}
                         }
                     }
                 }
-                if (backend.hostname !== 'screeps.com') {
+                if (!isOfficial) {
                     // Replace room-history URL
-                    text = text.replace(
+                    src = src.replace(
                         /http:\/\/"\+s\.options\.host\+":"\+s\.options\.port\+"\/room-history/g,
                         client.getURL(Route.HISTORY),
                     );
 
                     // Replace official CDN with local assets
-                    text = text.replace(/https:\/\/d3os7yery2usni\.cloudfront\.net/g, `${info.backend}/assets`);
+                    src = src.replace(/https:\/\/d3os7yery2usni\.cloudfront\.net/g, `${info.backend}/assets`);
                 }
 
-                // Replace URLs with proxy client root path
-                text = text.replace(/https:\/\/screeps.com\/a\//g, client.getURL(Route.ROOT));
+                // Replace URLs with local client paths
+                src = src.replace(/https:\/\/screeps.com\/a\//g, client.getURL(Route.ROOT));
             }
-            return argv.beautify ? jsBeautify(text) : text;
+            return argv.beautify ? jsBeautify(src) : src;
         } else {
             // JSZip doesn't implement their read stream correctly and it causes EPIPE crashes. Pass it
             // through a no-op transform stream first to iron that out.
@@ -402,7 +313,7 @@ koa.use(async (context, next) => {
         return;
     }
 
-    const info = extract(context.url);
+    const info = extractBackend(context.url, argv.backend);
     if (info) {
         context.respond = false;
         context.req.url = info.endpoint;
@@ -421,13 +332,13 @@ koa.use(async (context, next) => {
 
 // Proxy WebSocket requests
 server.on('upgrade', (req, socket, head) => {
-    const info = extract(req.url!);
+    const info = extractBackend(req.url!, argv.backend);
     if (info && req.headers.upgrade?.toLowerCase() === 'websocket') {
         req.url = info.endpoint;
         proxy.ws(req, socket, head, {
             target: argv.internal_backend ?? info.backend,
         });
-        socket.on('error', (err) => error(err));
+        socket.on('error', (err) => logError(err));
     } else {
         socket.end();
     }
