+++ src/utils/types.ts
@@ -1,5 +1,6 @@
 declare global {
     interface Window {
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
         angular: any;
     }
 }
